@@ -1,32 +1,68 @@
 {
-  "main": "node_modules/expo/AppEntry.js",
-  "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "eject": "expo eject",
-    "ios:beta": "fastlane ios beta",
-    "android:beta": "fastlane android beta"
+  "name": "restoic-web",
+  "version": "0.1.0",
+  "engines": {
+    "node": "v15.3.0"
   },
+  "private": true,
   "dependencies": {
-    "@expo/vector-icons": "^9.0.0",
-    "expo": "^32.0.0",
-    "expokit": "^32.0.3",
-    "lodash": "^4.17.13",
-    "native-base": "^2.10.0",
-    "prop-types": "^15.6.2",
-    "react": "16.5.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
-    "react-navigation": "^3.0.9",
-    "react-navigation-redux-helpers": "^2.0.9",
-    "react-redux": "^6.0.0",
-    "react-style-proptype": "^3.2.2",
-    "redux": "^4.0.1",
-    "redux-thunk": "^2.3.0"
+    "@testing-library/jest-dom": "^5.11.4",
+    "@testing-library/react": "^11.1.0",
+    "@testing-library/user-event": "^12.1.10",
+    "axios": "^0.21.1",
+    "connected-react-router": "^6.8.0",
+    "history": "4.10.1",
+    "lodash": "^4.17.21",
+    "react": "^17.0.1",
+    "react-apple-login": "^1.1.3",
+    "react-calendar": "^3.3.1",
+    "react-circular-progressbar": "^2.0.3",
+    "react-dom": "^17.0.1",
+    "react-google-login": "^5.2.2",
+    "react-html-parser": "^2.0.2",
+    "react-modal": "^3.12.1",
+    "react-redux": "^7.2.2",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "4.0.1",
+    "redux": "^4.0.5",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.3.0",
+    "web-vitals": "^0.2.4",
+    "workbox-background-sync": "^5.1.3",
+    "workbox-broadcast-update": "^5.1.3",
+    "workbox-cacheable-response": "^5.1.3",
+    "workbox-core": "^5.1.3",
+    "workbox-expiration": "^5.1.3",
+    "workbox-google-analytics": "^5.1.3",
+    "workbox-navigation-preload": "^5.1.3",
+    "workbox-precaching": "^5.1.3",
+    "workbox-range-requests": "^5.1.3",
+    "workbox-routing": "^5.1.3",
+    "workbox-strategies": "^5.1.3",
+    "workbox-streams": "^5.1.3"
   },
-  "devDependencies": {
-    "babel-preset-expo": "^5.0.0",
-    "expo-cli": "^2.10.1"
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
-  "private": true
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
